% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_public_holidays}
\alias{aus_public_holidays}
\title{Australian Public Holidays}
\format{A data frame with six variables: \code{tbc}}
\usage{
aus_public_holidays
}
\description{
This data was collected from the Australian Public Holidays page on australia.gov.au. It is being republished in a machine-readable
format on data.gov.au. The dataset is a single CSV covering all states and territories for 2019. This data is correct at the time of
publishing but is subject to change.
}
\keyword{datasets}
